move router-slot component
