+++ src/Umbraco.Web.UI.Client/src/packages/core/router/components/router-slot/index.ts
@@ -1,0 +1,5 @@
+export * from './route.context.js';
+export * from './router-slot-change.event.js';
+export * from './router-slot-init.event.js';
+export * from './router-slot.element.js';
+export type * from './route.interface.js';
