+++ src/Umbraco.Web.UI.Client/src/packages/core/router/components/router-slot/route.context.ts
@@ -1,6 +1,6 @@
+import { umbGenerateRoutePathBuilder } from '../../generate-route-path-builder.function.js';
+import type { UmbModalRouteRegistration } from '../../modal-registration/modal-route-registration.interface.js';
 import type { UmbRoute } from './route.interface.js';
-import { umbGenerateRoutePathBuilder } from './generate-route-path-builder.function.js';
-import type { UmbModalRouteRegistration } from './modal-registration/modal-route-registration.interface.js';
 import type { IRouterSlot } from '@umbraco-cms/backoffice/external/router-slot';
 import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
