+++ src/Umbraco.Web.UI.Client/src/packages/core/router/components/router-slot/route.interface.ts
