+++ src/Umbraco.Web.UI.Client/src/packages/core/router/components/router-slot/router-slot-change.event.ts
