+++ src/Umbraco.Web.UI.Client/src/packages/core/router/index.ts
@@ -1,12 +1,8 @@
-export * from '@umbraco-cms/backoffice/external/router-slot';
-export * from './encode-folder-name.function.js';
-export * from './route.context.js';
-export type * from './route.interface.js';
-export * from './router-slot-change.event.js';
-export * from './router-slot-init.event.js';
-export * from './router-slot.element.js';
-export * from './path-pattern.class.js';
-export type * from './modal-registration/modal-route-registration.interface.js';
-export * from './modal-registration/modal-route-registration.controller.js';
-export type * from './types.js';
 export * from './components/not-found/route-not-found.element.js';
+export * from './components/router-slot/index.js';
+export * from './encode-folder-name.function.js';
+export * from './modal-registration/modal-route-registration.controller.js';
+export * from './path-pattern.class.js';
+export * from '@umbraco-cms/backoffice/external/router-slot';
+export type * from './modal-registration/modal-route-registration.interface.js';
+export type * from './types.js';
